Endret antall i BarnIHusstand til Double og lagt til nytt grunnlagsobjekt BarnInfo

+++ src/main/kotlin/no/nav/bidrag/behandling/felles/grunnlag/BarnIHusstand.kt
@@ -2,7 +2,7 @@ package no.nav.bidrag.behandling.felles.grunnlag
 
 import java.time.LocalDate
 
-class BarnIHusstand(val antall: Int, datoFom: LocalDate, datoTil: LocalDate?) : Periode<BarnIHusstand>(datoFom, datoTil),
+class BarnIHusstand(val antall: Double, datoFom: LocalDate, datoTil: LocalDate?) : Periode<BarnIHusstand>(datoFom, datoTil),
   IGrunnlagInnhold {
 
   override fun isEqualTo(other: BarnIHusstand): Boolean {
